+++ oci_api/graph/zfs_filesystem.py
@@ -15,9 +15,15 @@
 import logging
 import pathlib
 import shutil
+import tarfile
+import tempfile
+import humanize
+
 from oci_api import OCIError, oci_config
-from oci_api.util.zfs import zfs_create, zfs_get, zfs_set, zfs_snapshot, \
-    zfs_send, zfs_destroy, zfs_is_snapshot, zfs_is_filesystem, zfs_list, zfs_clone
+from oci_api.util.zfs import (
+    zfs_create, zfs_get, zfs_set, zfs_snapshot, 
+    zfs_destroy, zfs_clone, zfs_diff
+)
 from oci_api.util.file import rm, untar, uncompress, du
 
 log = logging.getLogger(__name__)
@@ -32,12 +38,9 @@ class ZFSFilesystem:
         self.origin = origin
 
     @property
-    def size(self):
-        return du(self.path)
-
-    @property
     def path(self):
-        return pathlib.Path(oci_config['global']['path'], 'zfs', self.id)
+        if self.id is not None:
+            return pathlib.Path(oci_config['global']['path'], 'zfs', self.id)
 
     @property
     def zfs_filesystem(self):
@@ -48,16 +51,47 @@ class ZFSFilesystem:
     def zfs_snapshot(self):
         return '%s@diff' % self.zfs_filesystem
 
-    def zfs_size(self):
+    def size(self):
+        if self.path is not None:
+            return du(self.path)
+
+    def diff_size(self):
+        if self.id is None:
+            return None
+        log.debug('Start calculating diff size (%s)' % self.id)
+        origin_snapshot = None
+        if self.origin is not None:
+            origin_snapshot = self.origin.zfs_snapshot
+        path = self.path
+        size = 0
+        for change_info in zfs_diff(self.zfs_snapshot, origin_snapshot, include_file_types=True):
+            change_type = change_info[0]
+            file_type = change_info[1]
+            if (change_type == 'M' or change_type == '+') and file_type == 'F':
+                    file_path = pathlib.Path(change_info[2])
+                    file_size = file_path.stat().st_size
+                    log.debug('File %s, size %d' % (file_path, file_size))
+                    size += file_size
+        log.debug('Finish calculating diff size (%s)' % self.id)
+        return size
+
+    def virtual_size(self):
         # compressed and/or deduplicated size is smaller than actuall size
-         return zfs_get(self.zfs_filesystem, 'used')
+        return zfs_get(self.zfs_filesystem, 'used')
+
+    def zfs_filesystem_size(self):
+        # compressed and/or deduplicated size is smaller than actuall size
+        return zfs_get(self.zfs_filesystem, 'used')
 
     def create(self):
         if self.origin is None:
+            log.debug('Creating filesystem (%s)' % self.zfs_filesystem)
             zfs_filesystem = zfs_create(self.zfs_filesystem, mountpoint=self.path)
             if zfs_filesystem != self.zfs_filesystem:
                 raise OCIError('Could not create zfs filesystem (%s)' % self.zfs_filesystem)
         else:
+            log.debug('Cloning filesystem (%s) from (%s)' 
+                % (self.zfs_filesystem, self.origin.zfs_snapshot))
             zfs_filesystem = zfs_clone(self.zfs_filesystem, self.origin.zfs_snapshot, 
                 mountpoint=self.path)
             if zfs_filesystem != self.zfs_filesystem:
@@ -69,12 +103,47 @@ class ZFSFilesystem:
             raise OCIError('Could not destroy zfs filesystem (%s)' % self.zfs_filesystem)
         rm(self.path)
 
-    def add_tar_file(self, tar_file_path, destination_path=None):
-        local_destination_path = self.path
-        if destination_path is not None:
-            local_destination_path = local_destination_path.joinpath(destination_path).resolve()
-        untar(local_destination_path, tar_file_path=tar_file_path)
-
     def commit(self):
         zfs_set(self.zfs_filesystem, readonly=True)
         zfs_snapshot('diff', self.zfs_filesystem)
+    
+    def load_changeset(self, changeset_file_path):
+        log.debug('Start loading changeset (%s)' % str(changeset_file_path))
+        path = self.path
+        size = 0
+        with tarfile.open(changeset_file_path, "r") as tar_file:
+            for member in tar_file:
+                file_path = pathlib.Path(member.name)
+                if file_path.name.startswith('.wh.'):
+                    file_path = path.joinpath(file_path)
+                    if file_path.name == '.wh..wh..opq':
+                        rm(file_path.parent, recursive=True)
+                    else:
+                        file_path = file_path.parent.joinpath(file_path.name[4:])
+                        rm(file_path)
+                else:
+                    size += member.size
+                    tar_file.extract(member, path)
+        log.debug('Finish loading changeset (%s), size: %s' % 
+            (str(changeset_file_path), humanize.naturalsize(size)))
+    
+    def save_changeset(self, changeset_file_path):
+        log.debug('Start saving changeset (%s)' % str(changeset_file_path))
+        origin_snapshot = None
+        if self.origin is not None:
+            origin_snapshot = self.origin.zfs_snapshot
+        with tarfile.open(changeset_file_path, "w") as tar_file:
+            with tempfile.NamedTemporaryFile() as wh_temp_file:
+                path = self.path
+                for change_info in zfs_diff(self.zfs_snapshot, origin_snapshot):
+                    change_type = change_info[0]
+                    file_path = pathlib.Path(change_info[1])
+                    if change_type == 'M' or change_type == '+':
+                        tar_file.add(file_path, arcname=file_path.relative_to(path), recursive=False)
+                    elif change_type == '-' or change_type == 'R':
+                        file_path = file_path.parent.joinpath('.wh.' + file_path.name)
+                        tar_file.add(wh_temp_file.name, arcname=file_path.relative_to(path), recursive=False)
+                    if change_type == 'R':
+                        file_path = pathlib.Path(change_info[2])
+                        tar_file.add(file_path, arcname=file_path.relative_to(path), recursive=False)
+        log.debug('Finish saving changeset (%s)' % str(changeset_file_path))